refact:アカウント削除オーバーレイのコンポーネント化
